add the cache to bedrock as well
