+++ llm-service/app/services/models/providers/caii.py
@@ -44,6 +44,7 @@ from llama_index.core.llms import LLM
 from llama_index.core.postprocessor.types import BaseNodePostprocessor
 
 from ._model_provider import ModelProvider
+from ...utils import timed_lru_cache
 from ...caii.caii import (
     get_caii_llm_models,
     get_caii_embedding_models,
@@ -55,20 +56,6 @@ from ...caii.caii import (
 )
 from ...caii.types import ModelResponse
 from ...llama_utils import completion_to_prompt, messages_to_prompt
-
-def timed_lru_cache(seconds: int, maxsize: int = 128):
-    def wrapper_cache(func):
-        func = lru_cache(maxsize=maxsize)(func)
-        func.expiration = time.monotonic() + seconds
-
-        @functools.wraps(func)
-        def wrapped_func(*args, **kwargs):
-            if time.monotonic() >= func.expiration:
-                func.cache_clear()
-                func.expiration = time.monotonic() + seconds
-            return func(*args, **kwargs)
-        return wrapped_func
-    return wrapper_cache
 
 class CAIIModelProvider(ModelProvider):
     @staticmethod
